# packages/schemas/openapi.yaml
components:
  schemas:
...
    ExtractRequest:
      properties:
        description:
          anyOf:
            - maxLength: 2000
              type: string
            - type: 'null'
          title: Description
        image_url:
          maxLength: 12000000
          title: Image Url
          type: string
      required:
        - image_url
      title: ExtractRequest
      type: object
    ExtractResponse:
      properties:
        brand:
          $ref: '#/components/schemas/FieldInference'
        brand_color_hex:
          $ref: '#/components/schemas/FieldInference'
        category:
          $ref: '#/components/schemas/FieldInference'
        name:
          anyOf:
            - $ref: '#/components/schemas/FieldInference'
            - type: 'null'
        price:
          anyOf:
            - $ref: '#/components/schemas/FieldInference'
            - type: 'null'
        product_type:
          $ref: '#/components/schemas/FieldInference'
        selling_points:
          items:
            $ref: '#/components/schemas/FieldInference'
          title: Selling Points
          type: array
      required:
        - name
        - brand
        - category
        - product_type
        - price
        - brand_color_hex
        - selling_points
      title: ExtractResponse
      type: object
    FieldInference:
      properties:
        confidence:
          maximum: 1
          minimum: 0
          title: Confidence
          type: number
        reasoning:
          title: Reasoning
          type: string
        value:
          title: Value
      required:
        - value
        - confidence
        - reasoning
      title: FieldInference
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    KitListItem:
      properties:
        category:
          anyOf:
            - type: string
            - type: 'null'
          title: Category
        id:
          title: Id
          type: integer
        locale:
          anyOf:
            - type: string
            - type: 'null'
          title: Locale
        name:
          title: Name
          type: string
        name_en:
          anyOf:
            - type: string
            - type: 'null'
          title: Name En
        score:
          anyOf:
            - type: integer
            - type: 'null'
          title: Score
        sku:
          title: Sku
          type: string
        status:
          title: Status
          type: string
        thumbs:
          items:
            anyOf:
              - type: string
              - type: 'null'
          title: Thumbs
          type: array
        updated_at:
          anyOf:
            - type: string
            - type: 'null'
          title: Updated At
      required:
        - id
        - sku
        - name
        - name_en
        - status
        - score
        - locale
        - thumbs
      title: KitListItem
      type: object
    KitListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/KitListItem'
          title: Items
          type: array
        total:
          title: Total
          type: integer
      required:
        - items
        - total
      title: KitListResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
        - loc
        - msg
        - type
      title: ValidationError
      type: object
...
info:
  title: Viskit Studio API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/kits:
    get:
      description: |-
        Return kits joined with their product catalog row, paginated & filtered.

        ``thumbs`` is the concatenation of up-to-5 hero png_paths (slot 1..5) and
        up-to-9 detail png_paths (M1..M9) — 14 slots total, NULL-padded for any
        missing rows.  Callers render placeholder cells for NULL entries.

        ``recent`` is advisory; sort defaults to ``created_at DESC`` to preserve
        the EPIC-7 Dashboard call shape (``?recent=true&limit=6``).  Catalog
        (EPIC-8) passes ``offset``, ``status``, ``locale``, ``min_score``,
        ``category``, ``sort``, ``order`` for filtered/paginated views.
      operationId: list_kits_api_kits_get
      parameters:
        - in: query
          name: recent
          required: false
          schema:
            default: false
            title: Recent
            type: boolean
        - in: query
          name: limit
          required: false
          schema:
            default: 6
            maximum: 100
            minimum: 1
            title: Limit
            type: integer
        - in: query
          name: offset
          required: false
          schema:
            default: 0
            minimum: 0
            title: Offset
            type: integer
        - in: query
          name: status
          required: false
          schema:
            anyOf:
              - maxLength: 32
                type: string
              - type: 'null'
            title: Status
        - in: query
          name: locale
          required: false
          schema:
            anyOf:
              - maxLength: 8
                type: string
              - type: 'null'
            title: Locale
        - in: query
          name: min_score
          required: false
          schema:
            anyOf:
              - maximum: 100
                minimum: 0
                type: integer
              - type: 'null'
            title: Min Score
        - in: query
          name: category
          required: false
          schema:
            anyOf:
              - maxLength: 64
                type: string
              - type: 'null'
            title: Category
        - in: query
          name: sku
          required: false
          schema:
            anyOf:
              - maxLength: 64
                type: string
              - type: 'null'
            title: Sku
        - in: query
          name: sort
          required: false
          schema:
            default: created_at
            enum:
              - created_at
              - updated_at
              - score
            title: Sort
            type: string
        - in: query
          name: order
          required: false
          schema:
            default: desc
            enum:
              - asc
              - desc
            title: Order
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KitListResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Kits
      tags:
        - imagegen
...
  /api/kits/{kit_id}/extract:
    post:
      description: |-
        Extract per-field inferences from a product image.

        Uses the vision provider (registry role "vision"); falls back to "llm" if
        the vision role is unavailable (R2 mitigation).

        Reserved-prefix guard: POST to kit_id='_warmup' returns 404 — defensive
        against POST collision with the GET /_warmup/extract warmup endpoint.
      operationId: extract_api_kits__kit_id__extract_post
      parameters:
        - in: path
          name: kit_id
          required: true
          schema:
            title: Kit Id
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Extract
      tags:
        - extract
...